requirements.txt
streamlit==1.41.1
pandas==2.2.3
langchain-openai==0.2.4
python-dotenv==1.0.1
plotly==5.24.1
PyPDF2==4.0.2
tqdm==4.68.1
tiktoken>=0.7.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
